+++ Tests/FeatureProtocolTests.swift
@@ -12,14 +12,12 @@ final class FeatureProtocolTests: XCTestCase {
             productName: "name",
             productPrice: 10.0
         )
-        XCTAssertEqual(ratingBox.name, "RatingBox")
-        XCTAssertTrue(ratingBox.isActive)
-        XCTAssertEqual(ratingBox.callToAction, "Rate this product!")
-        XCTAssertEqual(ratingBox.actionButton, "Send Review")
 
-        XCTAssertEqual(ratingBox.productName, "name")
-        XCTAssertEqual(ratingBox.productPrice, 10.0)
-        XCTAssertNil(ratingBox.productDescription)
+        XCTAssertEqual(ratingBox.status, .unrequested)
+        XCTAssertEqual(ratingBox.name, "RatingBox")
+        XCTAssertEqual(ratingBox.args.productName, "name")
+        XCTAssertEqual(ratingBox.args.productPrice, 10.0)
+        XCTAssertNil(ratingBox.args.productDescription)
     }
 
     func test_id() {
@@ -54,7 +52,6 @@ final class FeatureProtocolTests: XCTestCase {
             productPrice: 10.0,
             productDescription: "description"
         )
-        ratingBox.impressionId = fakeImpressionId
 
         let expectedJSON: JSONObject = [
             "productName": "name",
@@ -62,7 +59,7 @@ final class FeatureProtocolTests: XCTestCase {
             "productDescription": "description"
         ]
 
-        XCTAssertEqual(try ratingBox.args(), expectedJSON)
+        XCTAssertEqual(try ratingBox.args.encodeToJSONObject(), expectedJSON)
     }
 
     func test_args_withNil() throws {
@@ -77,37 +74,93 @@ final class FeatureProtocolTests: XCTestCase {
             "productPrice": 10.0
         ]
 
-        XCTAssertEqual(try ratingBox.args(), expectedJSON)
+        XCTAssertEqual(try ratingBox.args.encodeToJSONObject(), expectedJSON)
     }
 
-    func test_update() throws {
+    func test_update_SHOULD_notUpdateWithOnRequest_invalidJSON() throws {
         let ratingBox = RatingBox(
             productName: "name",
             productPrice: 10.0,
             productDescription: "description"
         )
-        ratingBox.impressionId = fakeImpressionId
+        XCTAssertThrowsError(try ratingBox.update(request: .on(outputJson: JSONObject(), impressionId: nil)), "Invalid JSON should throw")
+        XCTAssertEqual(ratingBox.status, .unrequested)
+    }
 
-        XCTAssertThrowsError(try ratingBox.update(outputJson: JSONObject(), isActive: true), "Invalid JSON should throw")
+    func test_update_SHOULD_updateWithOnRequest_validJSON_nilImpressionId() throws {
+        let ratingBox = RatingBox(
+            productName: "name",
+            productPrice: 10.0,
+            productDescription: "description"
+        )
 
-        XCTAssertEqual(ratingBox.productName, "name")
-        XCTAssertEqual(ratingBox.productPrice, 10.0)
-        XCTAssertEqual(ratingBox.productDescription, "description")
-        XCTAssertEqual(ratingBox.impressionId, fakeImpressionId)
-        XCTAssertEqual(ratingBox.callToAction, "Rate this product!")
-        XCTAssertEqual(ratingBox.actionButton, "Send Review")
-
-        let newImpressionId = UUID().uuidString
         let updatedJSON: JSONObject = [
             "callToAction": "New Call",
             "actionButton": "New Button",
-            "_impressionId": newImpressionId
+            "_impressionId": "encode_impression_id"
         ]
 
-        try ratingBox.update(outputJson: updatedJSON, isActive: false)
-        XCTAssertEqual(ratingBox.impressionId, newImpressionId)
-        XCTAssertEqual(ratingBox.callToAction, "New Call")
-        XCTAssertEqual(ratingBox.actionButton, "New Button")
-        XCTAssertFalse(ratingBox.isActive)
+        try ratingBox.update(request: .on(outputJson: updatedJSON, impressionId: nil))
+        if case let .on(outputs) = ratingBox.status {
+            XCTAssertEqual(outputs.callToAction, "New Call")
+            XCTAssertEqual(outputs.actionButton, "New Button")
+            XCTAssertEqual(outputs._impressionId, "encode_impression_id")
+        } else {
+            XCTFail("expected a status of `on`.")
+        }
+    }
+
+    func test_update_SHOULD_updateWithOnRequest_validJSON_impressionId() throws {
+        let ratingBox = RatingBox(
+            productName: "name",
+            productPrice: 10.0,
+            productDescription: "description"
+        )
+
+        let updatedJSON: JSONObject = [
+            "callToAction": "New Call",
+            "actionButton": "New Button",
+            "_impressionId": "encoded_impression_id"
+        ]
+
+        try ratingBox.update(request: .on(outputJson: updatedJSON, impressionId: "new_impression_id"))
+        if case let .on(outputs) = ratingBox.status {
+            XCTAssertEqual(outputs.callToAction, "New Call")
+            XCTAssertEqual(outputs.actionButton, "New Button")
+            XCTAssertEqual(outputs._impressionId, "new_impression_id")
+        } else {
+            XCTFail("expected a status of `on`.")
+        }
+    }
+
+    func test_update_SHOULD_updateWithOffRequest() throws {
+        let ratingBox = RatingBox(
+            productName: "name",
+            productPrice: 10.0,
+            productDescription: "description"
+        )
+        try ratingBox.update(request: .off)
+        XCTAssertEqual(ratingBox.status, .off)
+    }
+
+    func test_event_SHOULD_bundleInputEventWithImpressionId() throws {
+        let ratingBox = RatingBox(
+            productName: "name",
+            productPrice: 10.0,
+            productDescription: "description"
+        )
+
+        let updatedJSON: JSONObject = [
+            "callToAction": "New Call",
+            "actionButton": "New Button",
+            "_impressionId": "encoded_impression_id"
+        ]
+
+        // Update the feature to have an active status
+        try ratingBox.update(request: .on(outputJson: updatedJSON, impressionId: "new_impression_id"))
+
+        let result = ratingBox.event(.rating(stars: 3))
+        XCTAssertEqual(try result?.event.serialized(), try RatingBox.Event.Rating(stars: 3).serialized())
+        XCTAssertEqual(result?.impressionId, "new_impression_id")
     }
 }