+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 The changelog for `CausalLabs/ios-client-sdk`. Also see the [releases](https://github.com/CausalLabs/ios-client-sdk/releases) on GitHub.
 
+## 0.10.0
+
+- Improved event signaling contract allowing for a better autocomplete experience.
+- Added a public protocol (`CausalClientProtocol`) that `CausalClient` conforms to allowing for mock instances of the client to be used in unit tests.
+- Added a public constructor to `CausalClient`.
+- Updated how a feature exposes arguments and outputs fetched from the server.
+  - Arguments are now available as an `args` variable on the feature object.
+  - Outputs are now available as a `status` variable on the feature. The status field is an enum which details the response returned from the impression server. The different status states are as follows:
+    - **unrequested**: The feature has been constructed but the client hasn't updated it yet.
+    - **on**: The feature was successfully loaded and is active (on). Access to the outputs is available as an associated value.
+    - **off**: The feature was successfully loaded and is not active (off).
+- Updated the compiler generated view models to expose a `state` object which helps the underlying view render all of the different states that the feature can be in. The different states are as follows:
+  - **on**: Indicates that the feature was successfully loaded (either from cache or network) and is active. Access to the feature outputs is available as an associated value.
+  - **off**: Indicates that the feature was successfully loaded (either from cache or network) and is not active. These should not be shown.
+  - **loading**: Indicates that the feature is currently being loaded.
+- Updated the `CausalClient.requestCacheFill` method to no longer mutate input arguments.
+
 ## 0.9.0
 
 - Feature caching fixes and improvements.